[feature/task1_5-6]: Renamed module-files and changed scripts in package.json and
