+++ package.json
@@ -3,11 +3,10 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng lint && ng serve -o",
+    "start": "npm run lint && ng serve -o",
     "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint",
-    "lint:fix": "ng lint --fix",
+    "lint": "ng lint --fix",
     "e2e": "ng e2e"
   },
   "private": true,
