+++ src/app/products/components/product/product.component.html
@@ -3,5 +3,5 @@
   <p>{{product.description}}</p>
   <p><b>Category:</b> {{product.category}}</p>
   <p><b>Price:</b> {{product.price | currency}}</p>
-  <button (click)="onBy()">Buy</button>
+  <button (click)="onBuy()">Buy</button>
 </div>
