+++ src/app/products/components/products-list/products-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Product } from '../../models/product';
+import { Product } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
 
 @Component({
