+++ src/app/products/models/product.model.ts
